# storage: use RangeList instead of RangeTree in CommandQueue