@nvanbenschoten 2017-11-23
I just ran some benchmarks on this change and it looks like it's a clear improvement:

```
name                                          old time/op    new time/op    delta
CommandQueueReadWriteMix/readsPerWrite=1-4      8.89ms ± 1%    6.35ms ± 1%  -28.59%  (p=0.000 n=8+9)
CommandQueueReadWriteMix/readsPerWrite=4-4      4.99ms ± 2%    3.95ms ± 1%  -20.88%  (p=0.000 n=9+9)
CommandQueueReadWriteMix/readsPerWrite=16-4     2.85ms ± 2%    2.65ms ± 2%   -6.90%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=64-4     1.57ms ± 3%    1.57ms ± 5%     ~     (p=0.481 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=128-4    1.32ms ± 1%    1.33ms ± 4%     ~     (p=0.165 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=256-4    1.20ms ± 1%    1.24ms ± 4%   +3.66%  (p=0.000 n=9+9)

name                                          old alloc/op   new alloc/op   delta
CommandQueueReadWriteMix/readsPerWrite=1-4      1.30MB ± 0%    0.74MB ± 0%  -43.31%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=4-4       992kB ± 0%     691kB ± 0%  -30.35%  (p=0.000 n=9+9)
CommandQueueReadWriteMix/readsPerWrite=16-4      790kB ± 0%     667kB ± 0%  -15.57%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=64-4      533kB ± 0%     501kB ± 0%   -5.94%  (p=0.000 n=10+9)
CommandQueueReadWriteMix/readsPerWrite=128-4     487kB ± 0%     472kB ± 0%   -3.12%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=256-4     464kB ± 0%     457kB ± 0%   -1.50%  (p=0.000 n=10+10)

name                                          old allocs/op  new allocs/op  delta
CommandQueueReadWriteMix/readsPerWrite=1-4       29.0k ± 0%      9.1k ± 0%  -68.69%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=4-4       16.2k ± 0%      8.2k ± 0%  -49.32%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=16-4      9.08k ± 0%     7.07k ± 0%  -22.12%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=64-4      5.40k ± 0%     4.88k ± 0%   -9.65%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=128-4     4.74k ± 0%     4.48k ± 0%   -5.35%  (p=0.000 n=10+10)
CommandQueueReadWriteMix/readsPerWrite=256-4     4.41k ± 0%     4.29k ± 0%   -2.73%  (p=0.000 n=10+8)
```