@tschottdorf 2017-11-23
These benchmarks look delicious, but I'm worried about scenarios with linear complexity that aren't covered by our coverage (I think we probably under-test high-contention scenarios in which the list size becomes significant).  Could you expand on what you expect for those? Have they been tested?